requirements.txt
# YOLO11 Training Pipeline Requirements
# Unilever Bangladesh Product Annotation Model

# Core dependencies
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0

# Data handling
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and logging
tqdm>=4.65.0

# Optional: For better performance
# albumentations>=1.3.0  # For advanced data augmentation
# tensorboard>=2.13.0     # For TensorBoard logging
# clearml>=1.13.0         # For experiment tracking
# wandb>=0.15.0           # For Weights & Biases logging

# Kaggle specific (uncomment if running on Kaggle)
# kaggle>=1.5.0
